assume that all "lights" are dimmable
Trying to infer whether an entity is dimmable based on the presence of
the "brightness" attribute is flaky, since lights in the "off" state
don't have a "brightness" attribute. This leads to differing behaviour
that's dependent upon the state of the lights at discovery time.

That's a shame. So, just assume that if it's a light, it's dimmable.
Non-dimmable lights might end up pretending they can be dimmed, but at
worst that's Alexa executing a nop, which is significantly less
frustrating than Alexa claiming that a light is not dimmable when you
know damn well that's a lie.